refactor: move some tab specific logic into tabs component
